Add modules to the functions
Fix linter errors
