+++ src/index.html
@@ -50,7 +50,7 @@
                     </tbody>
                     <tfoot>
                         <tr class="iAlign">
-                            <td class="col-12"><a href="#" class="btn btn-link custom-link">Clear all completed</a></td>
+                            <td class="col-12"><a id="btn" href="#" class="btn btn-link custom-link">Clear all completed</a></td>
                         </tr>
                     </tfoot>
                 </table>
